Add build script, config
Fix the RUNNER env var in the test script

Signed-off-by: Brian Cain <bcain@quicinc.com>

Use a runner to invoke qemu-hexagon for tests

Set the QEMU_LD_PREFIX

This was causing failures in test cases when they forked new processes.  The
children would invoke QEMU but without the context necessary to find the C
library.  The tests would fail with an error message like this:

    qemu-hexagon: Could not open '/lib/ld-musl-hexagon.so.1': No such file or directory

Current test results:

    $ rg '^test result' 16276_pass_138_fail.log
    151934:test result: ok. 16246 passed; 0 failed; 514 ignored; 0 measured; 0 filtered out; finished in 88.39s
    161336:test result: FAILED. 16276 passed; 138 failed; 346 ignored; 0 measured; 0 filtered out; finished in 500.53s

Signed-off-by: Brian Cain <bcain@quicinc.com>

cfg: fix the runner